requirements.txt

pyautogui


openai-agents


python-dotenv